Refactor: repo 구조 변경
